refactor: create auth module
