+++ src/common/types/index.ts
@@ -1,2 +1,1 @@
 export * from "./system";
-export * from "./review";
