+++ src/modules/auth/components/ConfirmSignUpNote.tsx
