+++ src/modules/reviews/components/ReviewItem/ReviewBody/ReviewBody.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/common/components/Button";
 import { reviewItemTheme, type ReviewItemVariants } from "../ReviewItem.theme";
-import { type Review } from "@/common/types";
+import { type Review } from "@/modules/reviews/types";
 
 export type ReviewBodyProps = ReviewItemVariants & {
   review: Review;
