+++ src/modules/reviews/components/ReviewItem/ReviewItem.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import { ReviewItem } from "./ReviewItem";
 import { UniversityAbbreviation } from "@prisma/client";
-import { type Review } from "@/common/types";
+import { type Review } from "@/modules/reviews/types";
 
 const review = {
   id: "1",
