+++ src/modules/reviews/components/ReviewItem/RevieweeGroup/RevieweeGroup.tsx
@@ -1,4 +1,4 @@
-import { type Review } from "@/common/types";
+import { type Review } from "@/modules/reviews/types";
 import { profileTheme } from "@/common/components/Profile";
 import { SchoolIcon } from "@/common/components/CustomIcon";
